Move js-tequilapi test sources to src
