+++ src/http/axios-adapter.spec.ts
@@ -17,8 +17,8 @@
 
 import axios from 'axios'
 import MockAdapter from 'axios-mock-adapter'
-import AxiosAdapter from '../../../src/adapters/axios-adapter'
-import TequilapiError from '../../../src/tequilapi-error'
+import AxiosAdapter from './axios-adapter'
+import TequilapiError from '../tequilapi-error'
 
 describe('TequilapiClient AxiosAdapter', () => {
   let adapter: AxiosAdapter
